+++ web/pages/newsletter/newsletter.routes.js
@@ -1,0 +1,10 @@
+import express from 'express'
+import {
+  getNewsletter,
+} from './newsletter.controller'
+
+const router = express.Router()
+
+router.get('/', getNewsletter)
+
+module.exports = router
