+++ web/pages/stats/stats.pug
@@ -1,8 +1,8 @@
-extends ./includes/layout
+extends ../../views/includes/layout
 
 block content
 
-	include ./includes/breadcrumb.pug
+	include ../../views/includes/breadcrumb.pug
 
 	if content
