+++ web/pages/stats/stats.routes.js
@@ -1,0 +1,10 @@
+import express from 'express'
+import {
+  getStats,
+} from './stats.controller'
+
+const router = express.Router()
+
+router.get('/', getStats)
+
+module.exports = router
