+++ web/pages/university/university.routes.js
@@ -1,0 +1,10 @@
+import express from 'express'
+import {
+  getUniversity,
+} from './university.controller'
+
+const router = express.Router()
+
+router.get('/', getUniversity)
+
+module.exports = router
