+++ web/web.routes.js
@@ -8,10 +8,9 @@ import portfolio from './pages/portfolio/portfolio.routes'
 import resume from './pages/resume/resume.routes'
 import talks from './pages/talks/talks.routes'
 import contact from './pages/contact/contact.routes'
-
-const stats = require('./routes/stats')
-const newsletter = require('./routes/newsletter')
-const university = require('./routes/university')
+import university from './pages/university/university.routes'
+import stats from './pages/stats/stats.routes'
+import newsletter from './pages/newsletter/newsletter.routes'
 
 const router = express.Router()
